Use echo as a router/middleware combination
`/*` -> It'll look for a static file, if it receives a `404` on the file search then it will return the React SPA.
`/api/*` -> This is where our backend handlers will be called, if a `404` is encountered here it'll be passed back to the user.

This also removes the need to have our own logging middleware and pass the logger around so much. Instead the logging middleware in `echo` is used and a logger is provided with the request context.